modif, note est décimale

+++ server/initialisation.sql
@@ -22,7 +22,7 @@ create table if not exists book (
     publisher varchar(255) not null,
     price_old varchar(255),
     price_new varchar(255),
-    rating int DEFAULT -1,
+    rating DECIMAL(2,1) DEFAULT -1,
     primary key (title, owner),
     constraint fk_book foreign key (owner) references user(email)
 );
